add container.py
