+++ warpfield/telescope/source.py
@@ -7,9 +7,10 @@ from astropy.coordinates import SkyCoord, Angle, Distance
 from astropy.table import QTable
 from astropy.time import Time
 from astroquery.gaia import Gaia
-import astropy.io.fits as fits
 import astropy.units as u
 import numpy as np
+
+from .container import QTableContainer
 
 
 __debug_mode__ = False
@@ -33,64 +34,6 @@ __columns__ = {
     'non_single_star': None,
     'ref_epoch': 'year',
 }
-
-
-@dataclass(frozen=True)
-class QTableContainer:
-    ''' QTable with I/O functions
-
-    Attributes:
-      table (QTable):
-          Table of celestial objects.
-    '''
-    table: QTable
-
-    def __len__(self):
-        return len(self.table)
-
-    def __getitem__(self, key):
-        return self.table[key]
-
-    def has(self, *items):
-        names = self.table.colnames
-        return all([name in names for name in items])
-
-    def get_dimension(self, name):
-        return u.get_physical_type(self.table[name])
-
-    @classmethod
-    def from_fitsfile(cls, filename, key='table'):
-        ''' Generate a SourceTable from a FITS file
-
-        Arguments:
-          filename (str):
-              The path to the source FITS file.
-          key (str):
-              The name of the FITS extension containing the table data.
-
-        Returns:
-          A table instance.
-        '''
-        hdul = fits.open(filename)
-        table = QTable.read(hdul[key])
-        return cls(table=table)
-
-    def writeto(self, filename, overwrite=False):
-        ''' Dump a SourceTable into a FITS file
-
-        Arguments:
-          filename (str):
-              The path to the output filename.
-
-        Options:
-          overwrite (bool):
-              An existing file will be overwritten if true.
-        '''
-        hdul = fits.HDUList([
-            fits.PrimaryHDU(),
-            fits.BinTableHDU(data=self.table, name='table')
-        ])
-        hdul.writeto(filename, overwrite=overwrite)
 
 
 def convert_skycoord_to_sourcetable(skycoord):
